// client/models/enums/stat_names.ts
export enum StatNames {
  SEED_QUANTITY = 'Seed Quantity',
  PLANT_QUALITY = 'Plant Quality',
  GROWING_TIME = 'Growing Time',
  PEST_RESISTANCE = 'Pest Resistance',
  DISEASE_RESISTANCE = 'Disease Resistance',
  NITROGEN_REQUIREMENT = 'Nitrogen Requirement',
  STEM_THICKNESS = 'Stem Thickness',
  STEM_HEIGHT = 'Stem Height',
  TEMP_TOLERANCE = 'Temperature Tolerance',
  MOIS_TOLERANCE = 'Moisture Tolerance'
}